# Conflicts:
#   frontend/src/pages/Favourites.tsx

+++ backend/src/controllers/gymController.ts
@@ -6,17 +6,35 @@ import cache from '../cache'
 import cloudinary from 'cloudinary'
 import GymAccount from '../models/GymAccount'
 import { IGymAccountWithId } from '@models/gymAccount'
+import axios from 'axios'
 
-const createGym = (req: Request, res: Response, next: NextFunction) => {
+const createGym = async (req: Request, res: Response, next: NextFunction) => {
   const { ctx } = req
   console.log('ctx: ', ctx)
   if (!ctx) return res.status(401).json({ error: 'Unauthorized' })
   const gymData = req.body
   const gymId = new mongoose.Types.ObjectId()
+  const coordinates = await getCoordinates(gymData.address)
+  const fullAddress = {
+    ...gymData.address,
+    location: coordinates
+      ? {
+          type: 'Point',
+          coordinates: coordinates,
+        }
+      : null,
+  }
+  console.log('CompleteAddress2:', fullAddress)
 
+  const cheapestOfferPrice =
+    gymData.offers.length > 0
+      ? Math.min(...gymData.offers.map((offer: any) => offer.priceEuro))
+      : 0
   const gym = new Gym({
     _id: gymId,
     ...gymData,
+    address: fullAddress,
+    cheapestOfferPrice,
   })
 
   return gym
@@ -29,7 +47,7 @@ const createGym = (req: Request, res: Response, next: NextFunction) => {
       res.status(201).json({ gym })
     })
     .catch((error) => {
-      console.error('Error saving gym:', error) // Detaillierte Fehlermeldung
+      console.error('Error saving gym:', error)
       res.status(500).json({ error: error.message })
     })
 }
@@ -50,7 +68,13 @@ const getGym = (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params
 
   return Gym.findById(id)
-    .then((gym) => res.status(200).json(gym))
+    .populate({
+      path: 'reviews.author',
+      select: 'displayName',
+    })
+    .then((gym) => {
+      res.status(200).json(gym)
+    })
     .catch((error) => res.status(500).json({ error }))
 }
 
@@ -75,6 +99,15 @@ const addReview = async (req: Request, res: Response) => {
     }
 
     gym.reviews.push(req.body.review)
+
+    const totalReviews = gym.reviews.length
+    const totalRating = gym.reviews.reduce(
+      (sum, review) => sum + review.rating.valueOf(),
+      0,
+    )
+
+    gym.averageRating = totalRating / totalReviews
+
     await gym.save()
     return res.status(201).json({ gym })
   } catch (err) {
@@ -82,36 +115,41 @@ const addReview = async (req: Request, res: Response) => {
   }
 }
 
-async function getCoordinates(
-  address: string,
+export async function getCoordinates(
+  address:
+    | {
+        street: string
+        postalCode: string
+        city: string
+        country: string
+      }
+    | string,
 ): Promise<[number, number] | void> {
+  let fullAddress: string
+  if (typeof address === 'string') {
+    fullAddress = address.toLowerCase()
+  } else {
+    fullAddress =
+      `${address.street}, ${address.postalCode} ${address.city}, ${address.country}`.toLowerCase()
+  }
   // Check if the coordinates are already cached to improve response time
-  const cachedCoordinates: [number, number] | undefined = cache.get(
-    address.toLowerCase(),
-  )
-
+  const cachedCoordinates: [number, number] | undefined = cache.get(fullAddress)
   if (cachedCoordinates !== undefined) return cachedCoordinates
 
-  return fetch(
-    `https://nominatim.openstreetmap.org/search?q=${address}&format=json`,
-  )
-    .then((response) => response.json())
-    .then((response: OpenStreetMapResponse[]) => {
-      if (!response || response.length === 0) {
-        return undefined
-      }
-      // We take the first result, they are ranked by importance
-      const result = response[0]
-      const coordinates = [parseFloat(result.lon), parseFloat(result.lat)] as [
-        number,
-        number,
-      ]
-      cache.set(address.toLowerCase(), coordinates)
-      return coordinates
-    })
-    .catch((error) => {
-      console.error('Error fetching coordinates:', error)
-    })
+  try {
+    const response = await axios.get(
+      `https://nominatim.openstreetmap.org/search?format=json&q=${encodeURIComponent(fullAddress)}`,
+    )
+    if (!response.data || response.data.length === 0) {
+      return
+    }
+    const { lat, lon } = response.data[0]
+    const coordinates = [parseFloat(lon), parseFloat(lat)] as [number, number]
+    cache.set(fullAddress, coordinates)
+    return coordinates
+  } catch (error) {
+    console.error('Error fetching coordinates:', error)
+  }
 }
 
 interface SearchRequestBody {
@@ -216,26 +254,38 @@ function gymBelongsToGymAccount(
   return gymIdsAsString.includes(gymId)
 }
 
-const deleteGym = (req: Request, res: Response, next: NextFunction) => {
+const deleteGym = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params
+  const { ctx } = req
 
-  if (!gymBelongsToGymAccount(req.ctx as IGymAccountWithId, id)) {
-    return res
-      .status(400)
-      .json({ message: "You don't have the right to delete this gym" })
+  if (!ctx) return res.status(401).json({ error: 'Unauthorized' })
+
+  try {
+    const gymAccountId = ctx._id
+    const gymAccount = await GymAccount.findById(gymAccountId)
+
+    if (!gymAccount) {
+      return res.status(404).json({ message: 'Gym account not found' })
+    }
+
+    if (!gymBelongsToGymAccount(req.ctx as IGymAccountWithId, id)) {
+      return res
+        .status(400)
+        .json({ message: "You don't have the right to delete this gym" })
+    }
+
+    const gym = await Gym.findByIdAndDelete(id)
+    if (!gym) {
+      return res.status(404).json({ message: 'Gym not found' })
+    }
+    gymAccount.gyms.pull(id) //mongoose method to remove entry from DocumentArray (ObjectIds)
+    await gymAccount.save()
+
+    return res.status(201).json({ message: 'Gym deleted' })
+  } catch (error) {
+    console.error('Error deleting gym:', error)
+    return res.status(500).json({ error })
   }
-
-  return Gym.findByIdAndDelete(id)
-    .then((gym) => {
-      if (gym) {
-        return res.status(201).json({ message: 'Gym deleted' })
-      } else {
-        return res.status(404).json({ message: 'Gym not found' })
-      }
-    })
-    .catch((error) => {
-      return res.status(500).json({ error })
-    })
 }
 
 const updateGym = (req: Request, res: Response, next: NextFunction) => {
